+++ __tests__/home.test.tsx
@@ -3,7 +3,7 @@
  */
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Home from '../app/page';
+import Home from '../app/padfgfd';
 
 describe('Home', () => {
   it('renders a main element', () => {
