--- app/(sections)/_hero/index.tsx
@@ -1,11 +1,0 @@
-import Background from "./background/background"
-
-
-export default function Hero() {
-
-  return (
-    <Background>
-    </Background>
-  )
-
-}
