+++ app/(sections)/about/page.tsx
@@ -1,0 +1,10 @@
+
+
+
+export default function Bot() {
+  return (
+    <section id="about" aria-label="about section">
+      about
+    </section>
+  )
+}
