--- app/(sections)/index.ts
@@ -1,5 +1,0 @@
-import Contact from "./contact/page";
-import Hero from "./_hero";
-import Projects from "./projects/page";
-
-export { Contact, Hero, Projects };
