+++ app/_utils/index.ts
@@ -1,3 +1,3 @@
 import hexToComplimentary from "./hexToComplimentary.js";
-
-export { hexToComplimentary }
+import { renderBackground, onResize } from "./three";
+export { hexToComplimentary, onResize, renderBackground }
